+++ src/librustc_mir/monomorphize/mod.rs
@@ -16,9 +16,11 @@ use rustc::ty::subst::Kind;
 use rustc::ty::{self, Ty, TyCtxt};
 
 pub use rustc::ty::Instance;
+pub use self::item::{MonoItem, MonoItemExt};
 
 pub mod collector;
 pub mod item;
+pub mod partitioning;
 
 fn fn_once_adapter_instance<'a, 'tcx>(
     tcx: TyCtxt<'a, 'tcx, 'tcx>,
